pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "connascence-analyzer"
version = "1.0.0"
description = "Professional connascence analysis for Python codebases"
authors = [{name = "Connascence Analytics", email = "hello@connascence.io"}]
license = {text = "BSL-1.1"}
readme = "docs/README.md"
keywords = ["code-quality", "coupling", "connascence", "static-analysis", "refactoring"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console", 
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "networkx>=2.8",
    "radon>=5.1.0",
    "click>=8.0.0",
    "rich>=12.0.0",
    "pathspec>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0", 
    "black>=22.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=2.20",
    "hypothesis>=6.0",
]
mcp = [
    "mcp>=1.0.0",  # When available
    "uvicorn>=0.20.0",
    "fastapi>=0.95.0",
]
vscode = [
    "websocket-client>=1.0.0",
]
enterprise = [
    "redis>=4.0.0",
    "sqlalchemy>=2.0.0", 
    "alembic>=1.8.0",
]

[project.urls]
Homepage = "https://connascence.io"
Repository = "https://github.com/connascence/connascence-analyzer"
Documentation = "https://docs.connascence.io"
Changelog = "https://github.com/connascence/connascence-analyzer/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/connascence/connascence-analyzer/issues"
"Commercial License" = "https://connascence.io/pricing"

[project.scripts]
connascence = "cli.connascence:main"

[project.entry-points."mcp.servers"]
connascence = "mcp.server:main"

# Tool Configuration

[tool.setuptools]
packages = [
    "analyzer", 
    "analyzer.ast_engine",
    "analyzer.dup_detection", 
    "analyzer.runtime_probe",
    "analyzer.frameworks",
    "policy",
    "integrators",
    "autofix", 
    "reporting",
    "cli",
    "mcp",
    "dashboard",
    "tooling",
]

[tool.setuptools.package-data]
policy = ["presets/*.yml", "baselines/*.json"]
reporting = ["templates/*.html", "templates/*.md"] 
vscode-extension = ["**/*"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | venv
  | build
  | dist
  | deprecated
  | experimental
)/
'''

[tool.ruff]
target-version = "py38" 
line-length = 120
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",  # raise ... from None
]
unfixable = [
    "B",  # Don't auto-fix potentially unsafe bugbear rules
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # asserts allowed in tests
    "ARG",     # unused function args allowed in tests
    "FBT",     # boolean trap allowed in tests
    "PLR2004", # magic values allowed in tests
]
"examples/**/*.py" = [
    "PLR2004", # magic values allowed in examples
    "T201",    # print allowed in examples
]

[tool.ruff.isort]
known-first-party = ["analyzer", "policy", "reporting", "cli", "mcp"]
force-sort-within-sections = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "networkx.*",
    "radon.*", 
    "pytest.*",
    "hypothesis.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--cov=analyzer",
    "--cov=policy",
    "--cov=reporting", 
    "--cov=cli",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests", "fixtures"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "property: marks tests as property-based tests",
]

[tool.coverage.run]
branch = true
source = ["analyzer", "policy", "reporting", "cli", "mcp"]
omit = [
    "tests/*",
    "fixtures/*",
    "examples/*", 
    "deprecated/*",
    "experimental/*",
    "*/conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstract",
]

[tool.coverage.html]
directory = "htmlcov"

# Connascence-specific configuration
[tool.connascence]
# Default policy for self-analysis
policy_preset = "strict-core"  

# Custom thresholds for this codebase
thresholds.max_positional_params = 4  # Slightly relaxed for internal APIs
thresholds.god_class_methods = 25     # Relaxed for main analyzer classes
thresholds.max_cyclomatic_complexity = 12

# Budget limits for PRs  
budgets.CoM = 8      # Magic literals
budgets.CoP = 5      # Position violations  
budgets.CoA = 3      # Algorithm duplication
budgets.critical = 0 # No critical issues
budgets.high = 5     # Max 5 high severity

# Exclusions specific to our project structure
exclusions = [
    "deprecated/*",
    "experimental/*", 
    "examples/*",
    "fixtures/*",
    "vscode-extension/out/*",
    "vscode-extension/node_modules/*",
]

# Framework profiles
frameworks = [
    "pytest",  # Test framework adaptations
    "fastapi", # For MCP server endpoints
    "vscode",  # VS Code extension patterns
]